frontend/requirements.txt
# Frontend Server Requirements
# Minimal dependencies for serving the frontend

# Web Framework
Flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0

# Environment Configuration
python-dotenv>=1.0.0
